NXBT-3716: Install and upgrade Cert Manager
To be compliant with Kubernetes 1.22.

The chart is installed in the cert-manager namespace, so it's a cluster-wide resource.
Yet, it is only installed with the "production" environment, aka when merging a pull request on the master branch.

Note: the chart was originally installed manually in version v1.11.1, see https://jira.nuxeo.com/browse/NXBT-3716.
Here it is upgraded to v1.12.0.